// src/BottomSheet/config/bottomSheet.types.ts
export type SnapPoint = {
  offset: number;
};